agents/assest_agent/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
crewai
crewai-tools
PyPDF2
pyyaml
pathlib
pandas
openpyxl
aiofiles
python-dotenv